Fixed plugin call which crashed with standard scripting

+++ PluginManager.h
@@ -70,7 +70,7 @@ typedef void (*ScopeSetIntGetTwoDbl)(int, double *, double*);
 typedef void (*ScopeSetIntGetIntDbl)(int, int *, double*);
 typedef int (*GetChannelList)(int, int, long *, const char **, BOOL, long *, long *,
                               BOOL *);
-typedef int (*LookupCamera)(const char *, BOOL, int, int *, float *, float *);
+typedef int (*LookupCamera)(const char *, BOOL, int, int *, double *, double *);
 typedef int (*FEIimage)(void *, int, int, int, double, double, int, int, int, long, long, 
                         long *, const char *, BOOL, int, BOOL, BOOL, long *);
 typedef int (*FEIchannels)(short **, const char **, int, long *, int, int, double,
